+++ src/components/page-layout/components/content.js
@@ -5,6 +5,8 @@ import classnames from 'classnames';
 import OnThisPage from '../../on-this-page/on-this-page';
 import Feedback from '../../feedback/feedback';
 import OverviewHeader from '../../overview-header/overview-header';
+import NextPage from './next-page.js';
+import GuideGroupIndex from './guide-group-index.js';
 
 export default class Content extends React.PureComponent {
   render() {
@@ -90,7 +92,8 @@ export class ContentWrapper extends React.PureComponent {
   };
 
   render() {
-    const { children, frontMatter, layoutConfig } = this.props;
+    const { children, frontMatter, layoutConfig, navigation, location } =
+      this.props;
     const { title, unProse, hideFeedback, layout, overviewHeader } =
       frontMatter;
     const { hideTitle } = layoutConfig;
@@ -122,7 +125,6 @@ export class ContentWrapper extends React.PureComponent {
               {this.renderAside(showFeedback)}
             </div>
           )}
-
           <div
             className={classnames('col', {
               'col--8-mxl col--12': layoutConfig.aside !== 'none',
@@ -130,6 +132,22 @@ export class ContentWrapper extends React.PureComponent {
             })}
           >
             {children}
+            {frontMatter.group && (
+              <GuideGroupIndex
+                pathname={location.pathname}
+                navigation={navigation}
+                frontMatter={frontMatter}
+              />
+            )}
+            {frontMatter.groupOrder && (
+              <div className="mt36">
+                <NextPage
+                  pathname={location.pathname}
+                  navigation={navigation}
+                  frontMatter={frontMatter}
+                />
+              </div>
+            )}
             {showFeedback && (
               <div
                 className={classnames('my36', {
@@ -155,7 +173,9 @@ ContentWrapper.propTypes = {
     headings: PropTypes.array,
     layout: PropTypes.string,
     overviewHeader: PropTypes.object,
-    onThisPage: PropTypes.bool
+    onThisPage: PropTypes.bool,
+    groupOrder: PropTypes.number,
+    group: PropTypes.bool
   }).isRequired,
   headings: PropTypes.array,
   location: PropTypes.object.isRequired,