+++ src/components/related-page/__tests__/related-page.test.js
@@ -153,4 +153,22 @@ describe('related-page', () => {
       expect(tree).toMatchSnapshot();
     });
   });
+
+  describe(testCases.next.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      testCase = testCases.next;
+      wrapper = renderer.create(
+        React.createElement(testCase.component, testCase.props)
+      );
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
 });